# swagger.yaml
swagger: '2.0'
info:
  description: >-
    This is the swagger documentation for Eirbmon Project in Enseirb-Matmeca
    2019/2020.
  version: 1.0.0
  title: Eirbmon
...
paths:
...
  /api/users:
    get:
      tags:
        - users
      summary: Retrieve all users
      operationId: GetAll
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ListUser'
    put:
      tags:
        - users
      summary: Update an user with id
      operationId: Update
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: User information that need to be updated
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Update'
        '404':
          description: User not found with id
  /api/auth:
    post:
      tags:
        - users
      summary: Authentification of an user
      operationId: Auth
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: username
          in: query
          description: Username of the user
          required: true
          type: string
        - name: password
          in: query
          description: Password of the user
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ReturnUser'
  /api/test:
    get:
      tags:
        - users
      summary: Test a token
      operationId: TestToken
      produces:
        - application/json
      parameters:
        - name: token
          in: path
          description: Token provided
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
        '403':
          description: Token wrong
  '/api/users/:_id':
    get:
      tags:
        - users
      summary: Retrieve a single user by id
      operationId: GetById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of the user
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
        '404':
          description: User not found by id
    delete:
      tags:
        - users
      summary: Delete an user by id
      operationId: Delete
      parameters:
        - name: id
          in: path
          description: The id of the user that needs to be deleted
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
        '404':
          description: User not found
        '500':
          description: Could not delete user
  '/api/users/email/:email':
    get:
      tags:
        - users
      summary: Retrieve a single user by email
      operationId: GetByEmail
      produces:
        - application/json
      parameters:
        - name: email
          in: path
          description: Email of the user
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
        '404':
          description: User not found by email
  '/api/users/name/:name':
    get:
      tags:
        - users
      summary: Retrieve a single user by username
      operationId: GetByUsername
      produces:
        - application/json
      parameters:
        - name: username
          in: path
          description: Username of the user
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: successful operation
        '404':
          description: User not found by username
  /api/users/send:
    post:
      tags:
        - users
      summary: Send an email to the user to reinitialize password
      operationId: SendMail
      parameters:
        - name: email
          in: path
          description: Email of the user
          required: true
          type: string
      responses:
        '200':
          description: OK
...
  '/api/eirbmon/name/:name':
    get:
      tags:
        - eirbmons
      summary: Retrieve a single Eirbmon by username
      operationId: GetByName
      produces:
        - application/json
      parameters:
        - name: name
          in: path
          description: Name of the Eirbmon
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: successful operation
        '404':
          description: Eirbmon not found by name
...
definitions:
...
  User:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
      owner_id:
        type: string
      permission:
        type: integer
        format: int32
    required:
      - username
      - email
      - password
  ReturnUser:
    type: object
    properties:
      token:
        type: integer
        format: int32
      username:
        type: string
      email:
        type: string
    required:
      - token
      - username
      - email
  ListUser:
    type: array
    items:
      type: object
      properties:
        schema:
          $ref: '#/definitions/User'
...
  Update:
    type: object
    properties:
      username:
        type: string
      email:
        type: string
      password:
        type: string
      owner_id:
        type: string
      permission:
        type: integer
        format: int32